+++ doc/Changelog.md
@@ -3,6 +3,9 @@ Changelog {#changelog}
 
 # git master {#master}
 
+* [#72](https://github.com/BlueBrain/Fivox/pull/72)
+  Replace TestLoader by GenericLoader, adding the possibility to load events
+  from file
 * [#69](https://github.com/BlueBrain/Fivox/pull/69)
   Add option in compute-vsd tool to export the VSD events as point sprite files
 * [#65](https://github.com/BlueBrain/Fivox/pull/65)